Improve conditional statement efficiency
Hi Daniel,

You've made good use of source control to save changes to your work on a regular basis.

Your code is pretty readable (good use of comments and generally good use of whitespace / blank lines) but in this pull request I've made some small suggestions to improve readability.

See my suggested change to the conditional statement – by connecting the branches of the conditional statement in this way, the efficiency of the code is increased. By using the "else-if" structure, the computer will only need to check *some* of the branches (rather than three separate IF statements).  The program will do the same thing, but your code is more efficient by using else-if structures.

Take care,

Mr. Gordon